desktop-shell: create a "managed_surface" interface
The new xdg-shell present/present_from_event() requests
specify that the compositor may notify the user. Thus,
we need a way to pass requests and events between the
compositor and the shell client (which does the drawing).

Create a "managed_surface" interface, which will keep track
of server-side shell surfaces, and is able to send requests
and receive events.

Signed-off-by: Manuel Bachmann <manuel.bachmann@iot.bzh>