desktop-shell: implement a present() GUI notifier
Each time a managed_surface receives a "presented"
event, a red tooltip will be created on the panel and
stacked near the clock.

The user can click on the tooltip to send an "activate"
from the shell to the compositor, which will bring the
surface to the foreground of the current workspace.

The tooltip will be destroyed after having been clicked,
or whenever the compositor-side shell surface gets
destroyed by any means.

Signed-off-by: Manuel Bachmann <manuel.bachmann@iot.bzh>