stacking: manage stacking lists, implement present()
We now keep track of each window's children, and thus are
able to close only these when a window gets destroyed.

If we try to create a new window and already have a child,
then we present() this child instead. A notification will
be issued and eventually allow us to raise the existing
child.

Signed-off-by: Manuel Bachmann <manuel.bachmann@iot.bzh>